Finalizando implementação do algoritmo e adicionando coluna de destino das arestas na database
