+++ main.py
@@ -5,32 +5,69 @@ class Graph:
 
 	def __init__(self, vertices):
 		self.V = vertices # qtd de vertices 
-		self.graph = []
+		self.graph = [] # Grafo
 
-	# Function to add an edge to graph
-	def addEdge(self, p1, lat1, lon1, p2, lat2, lon2):
+	def addEdge(self, tupla1, tupla2):
 		
 		# # calculando a distância entre os pontos usando latitude e longitude de cada aeroporto 
-		w = gdist((lat1, lon1), (lat2, lon2)).kilometers
+		w = gdist((tupla1[1], tupla1[2]), (tupla2[1], tupla2[2])).kilometers # peso da aresta = distancia em Km arredondados
 
-		self.graph.append([p1, p2, round(w)])
+		self.graph.append([tupla1[0], tupla2[0], round(w)])
 
+	
+	def find(self, parent, i):
+
+		if parent[i] != i:
+			parent[i] = self.find(parent, parent[i])
+
+		return parent[i]
+
+	
+	def union(self, parent, rank, x, y):
+
+		if rank[x] < rank[y]:
+			parent[x] = y
+
+		elif rank[x] > rank[y]:
+			parent[y] = x
+
+		else:
+			parent[y] = x
+			rank[x] += 1
 
 	def KruskalMST(self):
 
-		result = []
+		MST_result = []
 
-		# An index variable, used for sorted edges
+		# Iterador para as arestas ordenadas
 		i = 0
-
-		# An index variable, used for result[]
+		# Iterador para iterar a MST_result
 		e = 0
 
-		# ordena a lista do grafo pelo peso de cada aresta
+		# Ordena as arestas usando a chave como o peso de cada aresta
 		self.graph = sorted(self.graph, key=lambda item: item[2])
 
-		parent = []
+		parent = [] 
 		rank = []
+
+		for node in range(self.V):
+			parent.append(node)
+			rank.append(0)
+
+		while e < self.V - 1:
+			# armazena os valores da aresta de menor peso 
+			u, v, w = self.graph[i] 
+			i = i + 1
+
+			x = self.find(parent, u)
+			y = self.find(parent, v)
+
+			# Verifica se a aresta gera um ciclo
+			if x != y:
+				e = e + 1
+				MST_result.append([u, v, w])
+				self.union(parent, rank, x, y)
+
 
 # Leitura da base de dados
 db = pd.read_csv('airports.csv')
@@ -38,14 +75,17 @@ db = pd.read_csv('airports.csv')
 # Instância do grafo
 g = Graph(len(db))
 
-# # Captação dos dados da base de dados
-# for idx, row in db.iterrows():
-# 	print((row['AIRPORT'], row['LATITUDE'], row['LONGITUDE']))
+# Captação dos dados da base de dados e adição das arestas
+for idx, row in db.iterrows():
 
+	airport_1 = (idx , row['LATITUDE'], row['LONGITUDE'])
 
-g.addEdge('Yakutat', 59.50336056, -139.6602261, 'Yuma MCAS-Yuma International', 32.65658333, -114.6059722)
-g.addEdge('Eglin Air Force Base', 30.48325, -86.5254, 'Northwest Arkansas Regional', 36.28186944, -94.30681111)
-g.addEdge('Yellowstone', 44.68839917, -111.1176375, 'Yuma MCAS-Yuma International', 32.65658333, -114.6059722)
+	index_airport = row['Numero_Aleatorio'] # Index destino da Aresta
+	destiny_line = db.iloc[index_airport] # armazena o array da linha do Dataframe pelo index
+
+	airport_2 = (index_airport, destiny_line[5], destiny_line[6])
+
+	g.addEdge(airport_1, airport_2)
 
 
 g.KruskalMST()	
@@ -53,3 +93,18 @@ g.KruskalMST()
 print(g.graph)
 
 
+
+
+
+
+# PRINT A PARTIR DO INDICE DA BASE DE DADOS - pega os listas
+
+# import pandas as pd 
+
+# db = pd.read_csv('airports.csv')
+
+
+# indice_da_linha = 0
+# nome = db['AIRPORT'].iloc[indice_da_linha]
+
+# print(f'O nome na linha {indice_da_linha} é: {nome}')
